登录时查询用户权限和角色并加载到Spring Security的上下文中

+++ src/main/java/cn/xf/basedemo/mappers/SysPermissionMapper.java
@@ -15,6 +15,6 @@ import java.util.List;
 @Mapper
 public interface SysPermissionMapper extends BaseMapper<SysPermission> {
 
-	List<String> getPermissionListByRoleId(Long useId);
+	List<String> getPermissionListByRoleId(Integer useId);
 
 }
